requirements.txt
Flask==3.0.3
libmr==0.1.9
numpy==2.1.2
Pillow==11.0.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
